#30 feat : Hub_Route 구현 완료

+++ object/src/main/java/com/_P/eureka/client/object/hub/dto/HubResponseDto.java
@@ -14,9 +14,9 @@ public class HubResponseDto {
 
     private String address;
 
-    private int latitude;
+    private Long latitude;
 
-    private int longitude;
+    private Long longitude;
 
     public HubResponseDto(Hub hub){
         this.hubId = hub.getHubId();
